add party class update birthday class to inherit from party update dinner class to inherit from party form update to match correct base class props
