+++ partyPlanCalc/Form1.cs
@@ -19,17 +19,11 @@ namespace partyPlanCalc
         {
             InitializeComponent();
 
-            // initialize dinner party
-            // v1:
-            //dinnerParty = new DinnerParty() { NumberOfPeople = 5 };
-            //dinnerParty.SetHeathyOption(healthyBox.Checked);
-            //dinnerParty.CalculateCostOfDecorations(fancyBox.Checked);
-
-            // v2:
+            // initialize dinner party:
             dinnerParty = new DinnerParty((int)numericUpDown1.Value, healthyBox.Checked, fancyBox.Checked);
             DisplayDinnerPartyCost();
 
-            // initialize birthday party
+            // initialize birthday party:
             birthdayParty = new BirthdayParty((int)numberBirthday.Value, fancyBirthday.Checked, cakeWriting.Text);
             DisplayBirthdayPartyCost();
         }
@@ -37,11 +31,7 @@ namespace partyPlanCalc
         // dinner party methods
         private void DisplayDinnerPartyCost()
         {
-            // get cost of dinner party
-            // v1:
-            // decimal Cost = dinnerParty.CalculateCost(healthyBox.Checked);
-
-            // v2:
+            // get cost of dinner party:
             decimal Cost = dinnerParty.Cost;
 
             // display cost
@@ -50,38 +40,24 @@ namespace partyPlanCalc
 
         private void numericUpDown1_ValueChanged(object sender, EventArgs e)
         {
-            // set number of people
-            // v1:
-            // dinnerParty.NumberOfPeople = (int)numericUpDown1.Value;
-
-            // v2:
+            // set number of people:
             dinnerParty.NumberOfPeople = (int)numericUpDown1.Value;
 
-
+            // display cost
             DisplayDinnerPartyCost(); // always update cost label
         }
 
         private void fancyBox_CheckedChanged(object sender, EventArgs e)
         {
-            // set fancy option
-            // v1:
-            // dinnerParty.CalculateCostOfDecorations(fancyBox.Checked);
-
-            // v2:
-            dinnerParty.FancyDecoration = fancyBox.Checked;
-
+            // set fancy option:
+            dinnerParty.FancyDecorations = fancyBox.Checked;
             DisplayDinnerPartyCost(); // always update cost label
         }
 
         private void healthyBox_CheckedChanged(object sender, EventArgs e)
         {
-            // set healthy option
-            // v1:
-            // dinnerParty.SetHeathyOption(healthyBox.Checked);
-
-            // v2:
+            // set healthy option:
             dinnerParty.HealthyOption = healthyBox.Checked;
-
             DisplayDinnerPartyCost(); // always update cost label
         }
